incomplete work to try to guess the right mrconfig to register to
madduck suggested that mr register should register to the closest
mrconfig file. This branch is a start at trying to do that, by redefining
MR_CONFIG to hold either the config file that registers the given repo,
or for unregistered repos, the closest mrconfig file

Current issues:
- absolute path is always passed in MR_CONFIG, which makes mr config
  add new repos with absolute paths
- MR_CONFIG is not set to the closest mrconfig file for unregistered repos
  yet. The problem is, if I do this, then it will override the setting
  from -c, and so it won't be possible to use mr register to register in an
  explicitly specified config file.

I think it also has some other bugs. For example, the current mapping
between directory and config file only allows for one config file in a given
directory. Which might not be true if you're using mr -c.

As far as I can take it for now, so stashing away in this branch.