Fix fake-bare worktree checkout
We have to jump through hoops to check out fake-bare repositories since
the target directory probably already exists. Thus, we cheat by telling
git-clone *not* to checkout and then read-tree and checkout-index
ourselves. We also convert the non-bare repository to a fake bare one by
moving stuff around.

Probably a bit hackish but this is the best I can do for now.