Add an offline mode
When in offline mode, mr will remember commands that failed (presumably due
to lack of network access), and will run them again once back online. (This
is for all the laptop users like me who don't want to have to remember
which git repos to mr push when they come back from a nice long hack in the
woods. :-)