Makes hg push succeed if no changes must be pushed.
hg push fails with an exit code of 255 if there is an actual
error (see http://stackoverflow.com/questions/18536926).
This makes mr report hg push as a fail when it should not be.
This commit changes the invocation of hg push so that it
only really fails on exit code 255.