Work when HOME is a symlink. Closes: #793381
The reason this works is because $directory=getcwd(), which, when run
inside a symlinked directory, returns the path to the real directory.

So, $directory will be below the location that HOME is symlinked to,
and so it makes sense to call abs_path in the dirs from the mrconfig.

Based on a patch by Felix Eckhofer, but fixed to not throw undef when a
directory doesn't exist.