Add options for caching and printing cached output and return values
Cache each action and hook in separate directories so
that there is no confusion between them.

Use the full repo path as the basis of the cache path so
that it is simpler to calculate the location of the repo
based on the cache path.

This will be needed for the cache-mr-status.sh shell extension.