Use a pseudo-terminal when running commands whose output will be cached
The cached output will probably be printed on a terminal later and
the output for files, pipes and terminals will probably be the same
after stripping ANSI escape codes and caching both types of outputs
could be too expensive and complicated.

Strip ANSI escape codes when printing output to pipes or files.

This will be needed by users of the cache-mr-status.sh shell extension.

Regular-expressions-by: Adam Borowski <kilobyte@angband.pl>