Do not update the cache when in a subdirectory of the repository
The output of some repo commands is often different in subdirectories so
storing it in the cache would populate it with incorrect information.

The --top option should be used in subdirs of a repo to change dir to
the top of the repo and then update the cache.

This will be needed for the cache-mr-status.sh shell extension.