Add _default suffix for setting a fallback command for VCSen with no action
This is mostly useful for actions where the VCS-specific command for an action
should override the common command for an action rather than vice versa.
Currently, defining foo and VCS_foo will prefer foo over VCS_foo so this
additional fallback mechanism is needed in some situations.

This will be needed for the mr upgrade extension.