Add shell extension for incrementally updating the `mr status` output cache
This is most useful when there are a lot of repositories and the output
of `mr status` is accessed more often than the majority of the
repositories are changed.

When comparing daily diffs of the `mr --minimal status` output of a lot of
repositories, this should reduce the CPU and I/O usage and run-time a lot.

Input-from: martin f krafft <madduck@debian.org> (re $PWD & zsh)