Move cache-mr-status.sh out of the mr extension path
The documentation and examples encourage people to include all mr extensions
using a wildcard and that will cause errors when they upgrade to a version of
mr that includes the cache extension, which uses shell syntax not mr syntax.

Append .sh to the dir name since that is the customary extension for shell.